Move MergeTree*BloomFilter into the same header/source file
