--- src/Storages/MergeTree/MergeTreeIndexConditionBloomFilter.h
@@ -1,87 +1,0 @@
-#pragma once
-
-#include <Columns/IColumn.h>
-#include <Interpreters/BloomFilter.h>
-#include <Storages/MergeTree/KeyCondition.h>
-#include <Storages/MergeTree/MergeTreeIndices.h>
-#include <Storages/MergeTree/MergeTreeIndexGranuleBloomFilter.h>
-
-namespace DB
-{
-namespace ErrorCodes
-{
-    extern const int LOGICAL_ERROR;
-}
-
-class MergeTreeIndexConditionBloomFilter final : public IMergeTreeIndexCondition, WithContext
-{
-public:
-    struct RPNElement
-    {
-        enum Function
-        {
-            /// Atoms of a Boolean expression.
-            FUNCTION_EQUALS,
-            FUNCTION_NOT_EQUALS,
-            FUNCTION_HAS,
-            FUNCTION_HAS_ANY,
-            FUNCTION_HAS_ALL,
-            FUNCTION_IN,
-            FUNCTION_NOT_IN,
-            FUNCTION_UNKNOWN, /// Can take any value.
-            /// Operators of the logical expression.
-            FUNCTION_NOT,
-            FUNCTION_AND,
-            FUNCTION_OR,
-            /// Constants
-            ALWAYS_FALSE,
-            ALWAYS_TRUE,
-        };
-
-        RPNElement(Function function_ = FUNCTION_UNKNOWN) : function(function_) {} /// NOLINT
-
-        Function function = FUNCTION_UNKNOWN;
-        std::vector<std::pair<size_t, ColumnPtr>> predicate;
-    };
-
-    MergeTreeIndexConditionBloomFilter(const ActionsDAGPtr & filter_actions_dag, ContextPtr context_, const Block & header_, size_t hash_functions_);
-
-    bool alwaysUnknownOrTrue() const override;
-
-    bool mayBeTrueOnGranule(MergeTreeIndexGranulePtr granule) const override
-    {
-        if (const auto & bf_granule = typeid_cast<const MergeTreeIndexGranuleBloomFilter *>(granule.get()))
-            return mayBeTrueOnGranule(bf_granule);
-
-        throw Exception(ErrorCodes::LOGICAL_ERROR, "Requires bloom filter index granule.");
-    }
-
-private:
-    const Block & header;
-    const size_t hash_functions;
-    std::vector<RPNElement> rpn;
-
-    bool mayBeTrueOnGranule(const MergeTreeIndexGranuleBloomFilter * granule) const;
-
-    bool extractAtomFromTree(const RPNBuilderTreeNode & node, RPNElement & out);
-
-    bool traverseFunction(const RPNBuilderTreeNode & node, RPNElement & out, const RPNBuilderTreeNode * parent);
-
-    bool traverseTreeIn(
-        const String & function_name,
-        const RPNBuilderTreeNode & key_node,
-        const ConstSetPtr & prepared_set,
-        const DataTypePtr & type,
-        const ColumnPtr & column,
-        RPNElement & out);
-
-    bool traverseTreeEquals(
-        const String & function_name,
-        const RPNBuilderTreeNode & key_node,
-        const DataTypePtr & value_type,
-        const Field & value_field,
-        RPNElement & out,
-        const RPNBuilderTreeNode * parent);
-};
-
-}
